backend/requirements.txt
requests
feedparser
beautifulsoup4
pymongo
supabase
python-dotenv
google-generativeai
